backend/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
sqlalchemy==2.0.46
pydantic==2.12.5
python-dotenv==1.2.1
openai==2.21.0
PyJWT==2.11.0
passlib==1.7.4
bcrypt==5.0.0
Werkzeug==3.1.6
python-jose==3.5.0
psycopg2-binary==2.9.10
